+++ Postgres-cds_hub/init/344_cre_view_dataproc_core_last_import.sql
@@ -7,7 +7,7 @@
 -- Rights definition file size        : 16391 Byte
 --
 -- Create SQL Tables in Schema "db2dataprocessor_out"
--- Create time: 2025-11-11 12:32:58
+-- Create time: 2025-11-13 09:12:55
 -- TABLE_DESCRIPTION:  ./R-dataprocessor/dataprocessor/inst/extdata/Dataprocessor_Table_Description.xlsx[table_description]
 -- SCRIPTNAME:  344_cre_view_dataproc_core_last_import.sql
 -- TEMPLATE:  template_cre_view_last_import.sql
@@ -35,18 +35,50 @@ BEGIN
     ) THEN
 --------------------------------------------------------------------
 --Create View for frontend tables for schema db2dataprocessor_out
-DROP VIEW db2dataprocessor_out.v_input_data_files_last_import; -- first drop the view
 
-CREATE OR REPLACE VIEW db2dataprocessor_out.v_input_data_files_last_import AS (
-SELECT * FROM db_log.input_data_files
-WHERE TO_CHAR(COALESCE(last_check_datetime, input_datetime),'YYYY-MM-DD HH24:MI:SS') IN (SELECT TO_CHAR(MAX(COALESCE(last_check_datetime, input_datetime)),'YYYY-MM-DD HH24:MI:SS') FROM db_log.input_data_files)
-);
-DROP VIEW db2dataprocessor_out.v_input_data_files_processed_content_last_import; -- first drop the view
+DO
+$innerview$
+BEGIN
+    IF EXISTS ( -- do migration
+        SELECT 1 s FROM db_config.db_parameter WHERE parameter_name='current_migration_flag' AND parameter_value='1'
+    ) THEN
+        IF EXISTS ( -- VIEW exists
+            SELECT 1 s FROM information_schema.columns 
+            WHERE table_schema = 'db2dataprocessor_out' AND table_name = 'v_input_data_files_last_import'
+        ) THEN
+            DROP VIEW db2dataprocessor_out.v_input_data_files_last_import; -- first drop the view
+        END IF; -- DROP VIEW
+----------------------------
+        CREATE OR REPLACE VIEW db2dataprocessor_out.v_input_data_files_last_import AS (
+            SELECT * FROM db_log.input_data_files
+            WHERE TO_CHAR(COALESCE(last_check_datetime, input_datetime),'YYYY-MM-DD HH24:MI:SS') IN (SELECT TO_CHAR(MAX(COALESCE(last_check_datetime, input_datetime)),'YYYY-MM-DD HH24:MI:SS') FROM db_log.input_data_files)
+        );
+----------------------------
+    END IF; -- do migration
+END
+$innerview$;
 
-CREATE OR REPLACE VIEW db2dataprocessor_out.v_input_data_files_processed_content_last_import AS (
-SELECT * FROM db_log.input_data_files_processed_content
-WHERE TO_CHAR(COALESCE(last_check_datetime, input_datetime),'YYYY-MM-DD HH24:MI:SS') IN (SELECT TO_CHAR(MAX(COALESCE(last_check_datetime, input_datetime)),'YYYY-MM-DD HH24:MI:SS') FROM db_log.input_data_files_processed_content)
-);
+DO
+$innerview$
+BEGIN
+    IF EXISTS ( -- do migration
+        SELECT 1 s FROM db_config.db_parameter WHERE parameter_name='current_migration_flag' AND parameter_value='1'
+    ) THEN
+        IF EXISTS ( -- VIEW exists
+            SELECT 1 s FROM information_schema.columns 
+            WHERE table_schema = 'db2dataprocessor_out' AND table_name = 'v_input_data_files_processed_content_last_import'
+        ) THEN
+            DROP VIEW db2dataprocessor_out.v_input_data_files_processed_content_last_import; -- first drop the view
+        END IF; -- DROP VIEW
+----------------------------
+        CREATE OR REPLACE VIEW db2dataprocessor_out.v_input_data_files_processed_content_last_import AS (
+            SELECT * FROM db_log.input_data_files_processed_content
+            WHERE TO_CHAR(COALESCE(last_check_datetime, input_datetime),'YYYY-MM-DD HH24:MI:SS') IN (SELECT TO_CHAR(MAX(COALESCE(last_check_datetime, input_datetime)),'YYYY-MM-DD HH24:MI:SS') FROM db_log.input_data_files_processed_content)
+        );
+----------------------------
+    END IF; -- do migration
+END
+$innerview$;
 
 --SQL Role for Views in Schema db2dataprocessor_out
 GRANT SELECT ON TABLE db2dataprocessor_out.v_input_data_files_last_import TO db2dataprocessor_user;
