+++ Postgres-cds_hub/init/345_cre_view_dataproc_core_all.sql
@@ -7,7 +7,7 @@
 -- Rights definition file size        : 16391 Byte
 --
 -- Create SQL Tables in Schema "db2dataprocessor_out"
--- Create time: 2025-11-11 12:32:58
+-- Create time: 2025-11-13 09:12:55
 -- TABLE_DESCRIPTION:  ./R-dataprocessor/dataprocessor/inst/extdata/Dataprocessor_Table_Description.xlsx[table_description]
 -- SCRIPTNAME:  345_cre_view_dataproc_core_all.sql
 -- TEMPLATE:  template_cre_view_all.sql
@@ -35,21 +35,55 @@ BEGIN
     ) THEN
 --------------------------------------------------------------------
 --Create View for typed tables for schema db2dataprocessor_out
-------------------------------------------------------------------------------------------------------------------
--- sources are the plain typed data tables with a table name without any pre oder postfix -> SIMPLE_TABLE_NAME
-DROP VIEW db2dataprocessor_out.v_input_data_files; -- first drop the view
-CREATE OR REPLACE VIEW db2dataprocessor_out.v_input_data_files AS (SELECT * from db_log.input_data_files);
-
-GRANT SELECT ON db2dataprocessor_out.v_input_data_files TO db2dataprocessor_user;
-GRANT USAGE ON SCHEMA db2dataprocessor_out TO db2dataprocessor_user;
 
 ------------------------------------------------------------------------------------------------------------------
 -- sources are the plain typed data tables with a table name without any pre oder postfix -> SIMPLE_TABLE_NAME
-DROP VIEW db2dataprocessor_out.v_input_data_files_processed_content; -- first drop the view
-CREATE OR REPLACE VIEW db2dataprocessor_out.v_input_data_files_processed_content AS (SELECT * from db_log.input_data_files_processed_content);
+DO
+$innerview$
+BEGIN
+    IF EXISTS ( -- do migration
+        SELECT 1 s FROM db_config.db_parameter WHERE parameter_name='current_migration_flag' AND parameter_value='1'
+    ) THEN
+        IF EXISTS ( -- VIEW exists
+            SELECT 1 s FROM information_schema.columns 
+            WHERE table_schema = 'db2dataprocessor_out' AND table_name = 'v_input_data_files'
+        ) THEN
+            DROP VIEW db2dataprocessor_out.v_input_data_files; -- first drop the view
+        END IF; -- DROP VIEW
+----------------------------
+        CREATE OR REPLACE VIEW db2dataprocessor_out.v_input_data_files AS (SELECT * from db_log.input_data_files);
 
-GRANT SELECT ON db2dataprocessor_out.v_input_data_files_processed_content TO db2dataprocessor_user;
-GRANT USAGE ON SCHEMA db2dataprocessor_out TO db2dataprocessor_user;
+        GRANT SELECT ON db2dataprocessor_out.v_input_data_files TO db2dataprocessor_user;
+        GRANT USAGE ON SCHEMA db2dataprocessor_out TO db2dataprocessor_user;
+----------------------------
+    END IF; -- do migration
+END
+$innerview$;
+
+
+------------------------------------------------------------------------------------------------------------------
+-- sources are the plain typed data tables with a table name without any pre oder postfix -> SIMPLE_TABLE_NAME
+DO
+$innerview$
+BEGIN
+    IF EXISTS ( -- do migration
+        SELECT 1 s FROM db_config.db_parameter WHERE parameter_name='current_migration_flag' AND parameter_value='1'
+    ) THEN
+        IF EXISTS ( -- VIEW exists
+            SELECT 1 s FROM information_schema.columns 
+            WHERE table_schema = 'db2dataprocessor_out' AND table_name = 'v_input_data_files_processed_content'
+        ) THEN
+            DROP VIEW db2dataprocessor_out.v_input_data_files_processed_content; -- first drop the view
+        END IF; -- DROP VIEW
+----------------------------
+        CREATE OR REPLACE VIEW db2dataprocessor_out.v_input_data_files_processed_content AS (SELECT * from db_log.input_data_files_processed_content);
+
+        GRANT SELECT ON db2dataprocessor_out.v_input_data_files_processed_content TO db2dataprocessor_user;
+        GRANT USAGE ON SCHEMA db2dataprocessor_out TO db2dataprocessor_user;
+----------------------------
+    END IF; -- do migration
+END
+$innerview$;
 
 
 --------------------------------------------------------------------
